Migrate table from /docs/components/table to /components/table
- Port Basic and With Footer examples to the new reference page
- Remove old documentation page (site/ui/pages/table.tsx)
- Remove old route and import from routes.tsx
- Add Table entry to command palette
- Update E2E tests to target /components/table
