--- site/ui/pages/table.tsx
@@ -1,214 +1,0 @@
-/**
- * Table Documentation Page
- */
-
-import {
-  DocPage,
-  PageHeader,
-  Section,
-  Example,
-  PropsTable,
-  PackageManagerTabs,
-  type PropDefinition,
-  type TocItem,
-} from '../components/shared/docs'
-import { getNavLinks } from '../components/shared/PageNavigation'
-import { TablePreviewDemo, TableBasicDemo, TableWithFooterDemo } from '../components/table-demo'
-
-// Table of contents items
-const tocItems: TocItem[] = [
-  { id: 'installation', title: 'Installation' },
-  { id: 'examples', title: 'Examples' },
-  { id: 'basic', title: 'Basic', branch: 'start' },
-  { id: 'with-footer', title: 'With Footer', branch: 'end' },
-  { id: 'api-reference', title: 'API Reference' },
-]
-
-// Code examples
-const previewCode = `import {
-  Table, TableBody, TableCaption, TableCell,
-  TableFooter, TableHead, TableHeader, TableRow,
-} from '@/components/ui/table'
-
-const invoices = [
-  { invoice: 'INV001', status: 'Paid', method: 'Credit Card', amount: '$250.00' },
-  { invoice: 'INV002', status: 'Pending', method: 'PayPal', amount: '$150.00' },
-  { invoice: 'INV003', status: 'Unpaid', method: 'Bank Transfer', amount: '$350.00' },
-  { invoice: 'INV004', status: 'Paid', method: 'Credit Card', amount: '$450.00' },
-  { invoice: 'INV005', status: 'Paid', method: 'PayPal', amount: '$550.00' },
-  { invoice: 'INV006', status: 'Pending', method: 'Bank Transfer', amount: '$200.00' },
-  { invoice: 'INV007', status: 'Unpaid', method: 'Credit Card', amount: '$300.00' },
-]
-
-function TableDemo() {
-  return (
-    <Table>
-      <TableCaption>A list of your recent invoices.</TableCaption>
-      <TableHeader>
-        <TableRow>
-          <TableHead className="w-[100px]">Invoice</TableHead>
-          <TableHead>Status</TableHead>
-          <TableHead>Method</TableHead>
-          <TableHead className="text-right">Amount</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {invoices.map((inv) => (
-          <TableRow>
-            <TableCell className="font-medium">{inv.invoice}</TableCell>
-            <TableCell>{inv.status}</TableCell>
-            <TableCell>{inv.method}</TableCell>
-            <TableCell className="text-right">{inv.amount}</TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-      <TableFooter>
-        <TableRow>
-          <TableCell colSpan={3}>Total</TableCell>
-          <TableCell className="text-right">$2,500.00</TableCell>
-        </TableRow>
-      </TableFooter>
-    </Table>
-  )
-}`
-
-const basicCode = `import {
-  Table, TableBody, TableCell, TableHead, TableHeader, TableRow,
-} from '@/components/ui/table'
-
-const invoices = [
-  { invoice: 'INV001', status: 'Paid', method: 'Credit Card', amount: '$250.00' },
-  { invoice: 'INV002', status: 'Pending', method: 'PayPal', amount: '$150.00' },
-  { invoice: 'INV003', status: 'Unpaid', method: 'Bank Transfer', amount: '$350.00' },
-  { invoice: 'INV004', status: 'Paid', method: 'Credit Card', amount: '$450.00' },
-]
-
-function TableBasic() {
-  return (
-    <Table>
-      <TableHeader>
-        <TableRow>
-          <TableHead className="w-[100px]">Invoice</TableHead>
-          <TableHead>Status</TableHead>
-          <TableHead>Method</TableHead>
-          <TableHead className="text-right">Amount</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {invoices.map((inv) => (
-          <TableRow>
-            <TableCell className="font-medium">{inv.invoice}</TableCell>
-            <TableCell>{inv.status}</TableCell>
-            <TableCell>{inv.method}</TableCell>
-            <TableCell className="text-right">{inv.amount}</TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>
-  )
-}`
-
-const withFooterCode = `import {
-  Table, TableBody, TableCaption, TableCell,
-  TableFooter, TableHead, TableHeader, TableRow,
-} from '@/components/ui/table'
-
-const invoices = [
-  { invoice: 'INV001', status: 'Paid', method: 'Credit Card', amount: '$250.00' },
-  { invoice: 'INV002', status: 'Pending', method: 'PayPal', amount: '$150.00' },
-  { invoice: 'INV003', status: 'Unpaid', method: 'Bank Transfer', amount: '$350.00' },
-  { invoice: 'INV004', status: 'Paid', method: 'Credit Card', amount: '$450.00' },
-  { invoice: 'INV005', status: 'Paid', method: 'PayPal', amount: '$550.00' },
-  { invoice: 'INV006', status: 'Pending', method: 'Bank Transfer', amount: '$200.00' },
-  { invoice: 'INV007', status: 'Unpaid', method: 'Credit Card', amount: '$300.00' },
-]
-
-function TableWithFooter() {
-  return (
-    <Table>
-      <TableCaption>A list of your recent invoices.</TableCaption>
-      <TableHeader>
-        <TableRow>
-          <TableHead className="w-[100px]">Invoice</TableHead>
-          <TableHead>Status</TableHead>
-          <TableHead>Method</TableHead>
-          <TableHead className="text-right">Amount</TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {invoices.map((inv) => (
-          <TableRow>
-            <TableCell className="font-medium">{inv.invoice}</TableCell>
-            <TableCell>{inv.status}</TableCell>
-            <TableCell>{inv.method}</TableCell>
-            <TableCell className="text-right">{inv.amount}</TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-      <TableFooter>
-        <TableRow>
-          <TableCell colSpan={3}>Total</TableCell>
-          <TableCell className="text-right">$2,500.00</TableCell>
-        </TableRow>
-      </TableFooter>
-    </Table>
-  )
-}`
-
-// Props definitions
-const tableProps: PropDefinition[] = [
-  {
-    name: 'className',
-    type: 'string',
-    defaultValue: "''",
-    description: 'Additional CSS classes applied to the table element.',
-  },
-  {
-    name: 'children',
-    type: 'Child',
-    defaultValue: '-',
-    description: 'Table sub-components (TableHeader, TableBody, TableFooter, TableCaption).',
-  },
-]
-
-export function TablePage() {
-  return (
-    <DocPage slug="table" toc={tocItems}>
-      <div className="space-y-12">
-        <PageHeader
-          title="Table"
-          description="A responsive table component for displaying structured data."
-          {...getNavLinks('table')}
-        />
-
-        {/* Preview */}
-        <Example title="" code={previewCode}>
-          <TablePreviewDemo />
-        </Example>
-
-        {/* Installation */}
-        <Section id="installation" title="Installation">
-          <PackageManagerTabs command="barefoot add table" />
-        </Section>
-
-        {/* Examples */}
-        <Section id="examples" title="Examples">
-          <div className="space-y-8">
-            <Example title="Basic" code={basicCode}>
-              <TableBasicDemo />
-            </Example>
-
-            <Example title="With Footer" code={withFooterCode}>
-              <TableWithFooterDemo />
-            </Example>
-          </div>
-        </Section>
-
-        {/* API Reference */}
-        <Section id="api-reference" title="API Reference">
-          <PropsTable props={tableProps} />
-        </Section>
-      </div>
-    </DocPage>
-  )
-}
